Move test components to their own file.
