+++ test/support/component.ex
@@ -1,0 +1,10 @@
+defmodule Temple.Support.Component do
+  use Temple.Component
+
+  defmacro __using__(_) do
+    quote do
+      import Temple
+      import unquote(__MODULE__)
+    end
+  end
+end
